+++ recipes/default.rb
@@ -1,9 +1,15 @@
 #
-# Cookbook Name:: bcs_brightbox_ruby
+# Cookbook Name:: bcs_ruby
 # Recipe:: default
 #
-# Copyright (C) 2015 BCS Ltd.
-#
-# All rights reserved - Do Not Redistribute
-#
+# ruby dependencies
+%w(autoconf bison build-essential libssl-dev libyaml-dev libreadline6-dev
+   zlib1g-dev libncurses5-dev libffi-dev libgdbm3 libgdbm-dev).each do |pack|
+  package pack
+end
+
+node.default['brightbox-ruby']['version'] = node['ruby']['version']
+node.default['brightbox-ruby']['gems'] = node['ruby']['gems']
+node.default['brightbox-ruby']['install_ruby_switch'] = false
+
 include_recipe 'brightbox-ruby::default'