+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
+## Unreleased
+### breaking changes
+  - bcs_brightbox_ruby => bcs_ruby - no need to know which ruby installer I am using
+
 ## [1.0.1] - 2016-12-24
   - Rakefile added